msm8660: Perform typecast to disambiguate the call to abs in preparation for libcxx rebase

+++ legacy/msm8660/AudioHardware.cpp
@@ -2644,7 +2644,7 @@ uint32_t AudioHardware::getInputSampleRate(uint32_t sampleRate)
     uint32_t delta;
 
     for (i = 0, prevDelta = 0xFFFFFFFF; i < sizeof(inputSamplingRates)/sizeof(uint32_t); i++, prevDelta = delta) {
-        delta = abs(sampleRate - inputSamplingRates[i]);
+        delta = abs((int)(sampleRate - inputSamplingRates[i]));
         if (delta > prevDelta) break;
     }
     // i is always > 0 here
